Standardized variable size classes with struct members and return values across size_* functions.

+++ src/db.c
@@ -938,7 +938,7 @@ reverse(dbref list)
 size_t
 size_object(dbref i, int load)
 {
-    long byts;
+    size_t byts;
     byts = sizeof(struct object);
 
     if (NAME(i)) {
